scripts: imx-usb-loader: remove unnecessary check
file_base is the address where the image would be loaded to if we were
not skipping the unused space before the header in the image. I can't
see any reason why this should be higher than the download address.
Drop this superfluous check.

Tested on i.MX51, i.MX53 and i.MX6

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>