ARM: i.MX8M: add and use imx8mq_setup_pad()
We already have a mx8_setup_pad() function for early iomux setup, but it
is unused. Add a i.MX8MQ specific wrapper for the function which passes
the correct base address to mx8_setup_pad(). Let the boards use this
function. While at it rename mx8_setup_pad() to imx8_setup_pad() which
is more consistent to other i.MX specific functions.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>