ARM: i.MX: Drop iomux argument from mx7_setup_pad()
The name mx7_setup_pad already implies the SoC where it runs on, so we
do not have to pass the iomux base address but can hardcode it in the
function. While at it rename it to imx7_setup_pad() which is more
consistent to other i.MX specific functions.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>