ARM: i.MX8M: Add base addresses common to i.MX8M
This adds defines for the base addresses common to all currently
existing i.MX8M SoCs. Only the base addresses that are known to be
needed for the early SoC code are added. With this we can reuse the
early code for all variants without guessing that the base addresses
are the same for the other variants.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>