ARM: dts: phycore-imx6: change mtd-partition names for compatibility with kernel
The NVMEM-subsystem in newer kernels does not allow identical partition
names on different subdevices. According to the recommendation, we have to
rename the partition names to be compatible with actual kernel
versions.

To minimize the problems which could possibly arise, the nand device is
not changed. For all other devices, we prepend the device name to the
partitions.

Signed-off-by: Yunus Bas <y.bas@phytec.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>