net: Call net_poll() in a poller
This adds calling of net_poll() in a poller. With this we can react to
incoming packets like ping requests or fastboot requests.

We could change to call net_poll() from a poller exclusively, but this
would significantly slow down USB network controllers. As described in
the patch these take a long time in the packet receive path. To work
around this we keep the networking users call net_poll() at a high rate
when they are waiting for incoming packets and only every 10ms we call
net_poll() from a poller to get incoming traffic when no networking
protocol is actively calling net_poll().

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>