kbuild: remove dead code for modpost
rule_barebox-modpost is dead code since it is not invoked by anyone.

$(call barebox-modpost) is dead code too since there is no variable
named 'barebox-modpost'.

We could fix it to $(call rule_barebox-modpost) to invoke the rules,
but I hesitate to enable the code that has never been tested.

barebox-init, barebox-main are not set by anyone, so this code does
not seem to work.

The code is completely out of sync from the upstream Linux, so there
is no good reason to keep this rotten code.

If barebox needs modpost, this should be re-implemented correctly.
(preferably, by re-syncing with the latest Linux)

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>