reset: remove duplicate imx8mq-reset.h and use correct defines
The defines in the Barebox copy of imx8mq-reset.h had a typo, which has been fixed
in the upstream Linux version of this file. So depending on include ordering the
reset driver using those defines would fail to compile. Use the typo fixed versions
of the define in the driver and remove the duplicate include file, as we also have
the proper upstream version now.

Reported-by: Giorgio Dal Molin <giorgio.nicole@arcor.de>
Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>