nfs: Fix rpc_check_reply() return value for stale packets
When we receive a packet with the previous rpc_id then we have
the comment "stale packet, wait a bit longer", but that's not what
the code does. rpc_check_reply() returns 0 in this case and the caller
then interprets the packet as valid.
Always return -EAGAIN for invalid rpc_ids. This lets the caller ignore
the packets as intented.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>