mfd: double-quote the menu prompt of MFD
Linux commit 801b27db4638 ("kconfig: drop T_WORD from the RHS of
'prompt' symbol") requires every prompt to be quoted even if it is
a single word.

This is the only place with no quoting. It must be surrounded by
double-quotes in order to sync Kconfig to Linux 5.6-rc1 or later.

While I was here, I reworded "MFD" to "Multifunction device drivers"
to be consistent with Linux.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>