image: support 'powerpc' for mkimage architecture
Currently, the -A option of mkimage is determined based on $(SRCARCH).

I want to rename arch/ppc/ to arch/powerpc/.

Support 'powerpc' for the -A option of mkimage beforehand, otherwise
the barebox.uimage build would be broken.

common/image.c in U-Boot supports both 'powerpc' and 'ppc' for the
name of IH_ARCH_PPC. Let's do likewise in barebox.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>