net: cpsw: cpsw_process should only handle rx channels for its own port
The driver is currently processing the whole channel list, so it is
possible that it handles a channel of the wrong interface. We limit the
rx_chan handling for its desired port.

The cpsw_send is removing finished tx_chan transfers on each send, so this
has no directional limitation.

Signed-off-by: Michael Grzeschik <m.grzeschik@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>