mci: stm32_sdmmc2: move command timeout errors to debug level
Command timeouts are sometimes a normal operating condition
(for example when probing whether the connected card is SD or MMC),
so we don't want to print an error message for all of them. Move
those messages to the debug level, the MCI core will print some
more meaningful messages on most command errors anyways.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>